requirements.txt
# Core dependencies
mlx>=0.0.10
numpy>=1.24.0
pillow>=10.0.0
torch>=2.1.0  # Still needed for VAE decoding and video saving
torchvision>=0.16.0  # For video saving utilities

# Model loading and processing
transformers>=4.36.0  # For text encoders
huggingface_hub>=0.20.0  # For model downloads
safetensors>=0.4.0  # For CLIP model loading
python-dotenv>=1.0.0  # For .env handling

# UI and visualization
gradio>=4.0.0

# Utilities
loguru>=0.7.0
psutil>=5.9.0
tqdm>=4.66.0

# Optional dependencies
moviepy>=1.0.3  # For video processing
imageio>=2.31.0  # For image/video I/O
